+++ RELEASE_NOTES.md
@@ -2,7 +2,7 @@ Release Notes
 =============
 
 ## 1.9.15
-
+* Container Apps: Adds health probes.
 * AKS: Add node_taints parameter to agent pool config.
 
 ## 1.9.14